Used react bootstraps to create myNavbar
